+++ src/Model/Facility_Type.php
@@ -16,7 +16,12 @@ class Facility_Type extends Base
 
     public function name()
     {
-        return $this->_model->language;
+        return $this->_model->translations->name;
+    }
+
+    public function language()
+    {
+        return $this->_model->translations->language;
     }
 
     public function raw_data()